implement NotifySubscriptionsChangedRequestSubscriber add NotifySubscriptionsBuilder add metadata to notify config
