+++ Sources/WalletConnectNotify/Client/Wallet/ProtocolEngine/wc_notifySubscriptionsChanged/NotifySubscriptionsChangedRequestSubscriber.swift
@@ -8,19 +8,19 @@ class NotifySubscriptionsChangedRequestSubscriber {
     private var publishers = [AnyCancellable]()
     private let logger: ConsoleLogging
     private let notifyStorage: NotifyStorage
-    private let subscriptionScopeProvider: SubscriptionScopeProvider
+    private let notifySubscriptionsBuilder: NotifySubscriptionsBuilder
 
     init(networkingInteractor: NetworkInteracting,
          kms: KeyManagementServiceProtocol,
          logger: ConsoleLogging,
          notifyStorage: NotifyStorage,
-         subscriptionScopeProvider: SubscriptionScopeProvider
+         notifySubscriptionsBuilder: NotifySubscriptionsBuilder
     ) {
         self.networkingInteractor = networkingInteractor
         self.kms = kms
         self.logger = logger
         self.notifyStorage = notifyStorage
-        self.subscriptionScopeProvider = subscriptionScopeProvider
+        self.notifySubscriptionsBuilder = notifySubscriptionsBuilder
         subscribeForNofifyChangedRequests()
     }
 
@@ -29,7 +29,32 @@ class NotifySubscriptionsChangedRequestSubscriber {
         let protocolMethod =  NotifySubscriptionsChangedRequest()
 
         networkingInteractor.requestSubscription(on: protocolMethod).sink { [unowned self]  (payload: RequestSubscriptionPayload<NotifySubscriptionsChangedRequestPayload.Wrapper>) in
-            <#code#>
-        }
+
+
+            Task(priority: .high) {
+                logger.debug("Received Subscriptions Changed Request")
+
+                guard
+                    let (responsePayload, _) = try? NotifySubscriptionsChangedRequestPayload.decodeAndVerify(from: payload.request)
+                else { fatalError() /* TODO: Handle error */ }
+
+                // todo varify signature with notify server diddoc authentication key
+
+                let subscriptions = try await notifySubscriptionsBuilder.buildSubscriptions(responsePayload.subscriptions)
+
+                notifyStorage.replaceAllSubscriptions(subscriptions)
+
+                var logProperties = [String: String]()
+                for (index, subscription) in subscriptions.enumerated() {
+                    let key = "subscription_\(index + 1)"
+                    logProperties[key] = subscription.topic
+                }
+
+                logger.debug("Updated Subscriptions by Subscriptions Changed Request", properties: logProperties)
+
+            }
+
+        }.store(in: &publishers)
     }
+
 }
