+++ Sources/WalletConnectNotify/Client/Wallet/ProtocolEngine/wc_notifyWatchSubscriptions/NotifyWatchSubscriptionsResponseSubscriber.swift
@@ -7,19 +7,19 @@ class NotifyWatchSubscriptionsResponseSubscriber {
     private var publishers = [AnyCancellable]()
     private let logger: ConsoleLogging
     private let notifyStorage: NotifyStorage
-    private let subscriptionScopeProvider: SubscriptionScopeProvider
+    private let notifySubscriptionsBuilder: NotifySubscriptionsBuilder
 
     init(networkingInteractor: NetworkInteracting,
          kms: KeyManagementServiceProtocol,
          logger: ConsoleLogging,
          notifyStorage: NotifyStorage,
-         subscriptionScopeProvider: SubscriptionScopeProvider
+         notifySubscriptionsBuilder: NotifySubscriptionsBuilder
     ) {
         self.networkingInteractor = networkingInteractor
         self.kms = kms
         self.logger = logger
         self.notifyStorage = notifyStorage
-        self.subscriptionScopeProvider = subscriptionScopeProvider
+        self.notifySubscriptionsBuilder = notifySubscriptionsBuilder
         subscribeForWatchSubscriptionsResponse()
     }
 
@@ -38,7 +38,7 @@ class NotifyWatchSubscriptionsResponseSubscriber {
 
                     // todo varify signature with notify server diddoc authentication key
 
-                    let subscriptions = try await buildSubscriptions(responsePayload.subscriptions)
+                    let subscriptions = try await notifySubscriptionsBuilder.buildSubscriptions(responsePayload.subscriptions)
 
                     notifyStorage.replaceAllSubscriptions(subscriptions)
 
@@ -54,32 +54,4 @@ class NotifyWatchSubscriptionsResponseSubscriber {
             }.store(in: &publishers)
     }
 
-    private func buildSubscriptions(_ notifyServerSubscriptions: [NotifyServerSubscription]) async throws -> [NotifySubscription] {
-        var result = [NotifySubscription]()
-
-        for subscription in notifyServerSubscriptions {
-            let scope = try await buildScope(selectedScope: subscription.scope, dappUrl: subscription.dappUrl)
-            guard let metadata = try? await subscriptionScopeProvider.getMetadata(dappUrl: subscription.dappUrl),
-                  let topic = try? SymmetricKey(hex: subscription.symKey).derivedTopic() else { continue }
-
-            let notifySubscription = NotifySubscription(topic: topic,
-                                                        account: subscription.account,
-                                                        relay: RelayProtocolOptions(protocol: "irn", data: nil),
-                                                        metadata: metadata,
-                                                        scope: scope,
-                                                        expiry: subscription.expiry,
-                                                        symKey: subscription.symKey)
-            result.append(notifySubscription)
-        }
-
-        return result
-    }
-
-
-    private func buildScope(selectedScope: [String], dappUrl: String) async throws -> [String: ScopeValue] {
-        let availableScope = try await subscriptionScopeProvider.getSubscriptionScope(dappUrl: dappUrl)
-        return availableScope.reduce(into: [:]) {
-            $0[$1.name] = ScopeValue(description: $1.description, enabled: selectedScope.contains($1.name))
-        }
-    }
 }
