+++ Sources/WalletConnectNotify/Types/DataStructures/NotificationConfig.swift
@@ -5,4 +5,5 @@ struct NotificationConfig: Codable {
     let version: Int
     let lastModified: TimeInterval
     let types: [NotificationType]
+    let metadata: AppMetadata
 }
